feat(line): move stories to the dedicated storybook workspace
